Remove unnecessary file for test case
